get route for editanswer and edit button

+++ api/routes/editAnswersRouter/editAnswers.respository.js
@@ -3,6 +3,12 @@ module.exports = {
   editAnswer: async (answerId, starFlag, isStarred, isReviewed) => {
     try {
       const db = await get_db();
+      const answer = await db.query(`SELECT * FROM answers WHERE id = $1`, [
+        answerId,
+      ]);
+      if (!answer.rows[0]) {
+        return false;
+      }
       if (starFlag) {
         await db.query(
           `UPDATE answers 
@@ -18,6 +24,7 @@ module.exports = {
           [isReviewed, answerId]
         );
       }
+      return true;
     } catch (error) {
       throw Error(error);
     }
@@ -25,7 +32,24 @@ module.exports = {
   deleteAnswer: async (answerId) => {
     try {
       const db = await get_db();
-      await db.query(`DELETE FROM answers WHERE id = $1`, [answerId]);
+      const answer = await db.query(`SELECT * FROM answers WHERE id = $1`, [
+        answerId,
+      ]);
+      if (answer.rows[0]) {
+        await db.query(`DELETE FROM answers WHERE id = $1`, [answerId]);
+        return true;
+      } else return false;
+    } catch (error) {
+      throw Error(error);
+    }
+  },
+  getAnswer: async (answerId) => {
+    try {
+      const db = await get_db();
+      const result = await db.query(`SELECT * FROM answers WHERE id = $1`, [
+        answerId,
+      ]);
+      return result.rows[0];
     } catch (error) {
       throw Error(error);
     }
